requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
librosa==0.10.1
scikit-learn==1.3.0
pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
dash==2.14.1
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
face-recognition==1.3.0
dlib>=19.24.0
mediapipe==0.10.3
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2 